# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddressResponse:
    properties:
      address_id:
        type: string
      city:
        type: string
      country:
        type: string
      state:
        type: string
      street:
        type: string
      zip_code:
        type: string
    type: object
  dto.CreateAddressRequest:
    properties:
      city:
        type: string
      country:
        description: ISO 3166-1 alpha-2 e.g. "BR"
        type: string
      state:
        type: string
      street:
        type: string
      zip_code:
        type: string
    required:
    - city
    - country
    - state
    - street
    - zip_code
    type: object
...
  dto.CreateMerchantRepRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 2
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        enum:
        - admin
        - manager
        - staff
        type: string
    required:
    - email
    - name
    - password
    - phone
    - role
    type: object
...
  dto.CreateProfileRequest:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/dto.CreateAddressRequest'
        description: ← nested
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    required:
    - address
    - first_name
    - last_name
    - phone_number
    type: object
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  dto.MerchantRepResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      merchant_id:
        type: string
      merchant_rep_id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
...
  dto.PaginatedUsersResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.ProfileResponse:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/dto.AddressResponse'
        description: ← nested
      created_at:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      profile_id:
        type: string
      user_id:
        type: string
    type: object
  dto.UpdateAddressRequest:
    properties:
      city:
        type: string
      country:
        type: string
      state:
        type: string
      street:
        type: string
      zip_code:
        type: string
    type: object
...
  dto.UpdateMerchantRepRequest:
    properties:
      name:
        maxLength: 100
        minLength: 2
        type: string
      phone:
        type: string
      role:
        enum:
        - admin
        - manager
        - staff
        type: string
    type: object
...
  dto.UpdateProfileRequest:
    properties:
      address:
        allOf:
        - $ref: '#/definitions/dto.UpdateAddressRequest'
        description: ← nested
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
    type: object
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
host: ticketfair.localhost
info:
  contact:
    email: support@ticketfair.com
    name: TicketFair Support
  description: TicketFair event ticketing platform API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: TicketFair API
  version: "1.0"
paths:
...
  /merchant/rep/{id}:
    put:
      consumes:
      - application/json
      description: Update a rep's details under the authenticated merchant
      parameters:
      - description: Merchant Rep ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated rep data
        in: body
        name: merchantRep
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateMerchantRepRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MerchantRepResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Update a merchant representative
      tags:
      - Merchant Representatives
  /merchant/rep/new:
    post:
      consumes:
      - application/json
      description: Create a new rep under the authenticated merchant
      parameters:
      - description: Merchant rep data
        in: body
        name: merchantRep
        required: true
        schema:
          $ref: '#/definitions/dto.CreateMerchantRepRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MerchantRepResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Create a new merchant representative
      tags:
      - Merchant Representatives
...
  /private/profile:
    get:
      description: Retrieve the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get the current user's profile
      tags:
      - Profile
    post:
      consumes:
      - application/json
      description: Create a profile for the authenticated user
      parameters:
      - description: Profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Create a user profile
      tags:
      - Profile
    put:
      consumes:
      - application/json
      description: Update the authenticated user's profile
      parameters:
      - description: Updated profile data
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Update a user profile
      tags:
      - Profile
  /private/users:
    get:
      description: Retrieve a paginated list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedUsersResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all users
      tags:
      - Users
  /private/users/{id}:
    get:
      description: Retrieve a user by their UUID
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user by ID
      tags:
      - Users
  /private/users/me:
    get:
      description: Retrieve profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get the current authenticated user
      tags:
      - Users
...
  /public/auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user with email, password, and username
      parameters:
      - description: User registration data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"